genai_research_scientist/requirements.txt
streamlit>=1.37.0
scikit-learn>=1.4.0
numpy>=1.26.0
reportlab>=4.0.0
litellm>=1.43.0
python-dotenv>=1.0.1
requests>=2.32.0


